Refactor/ Header분리, 중복컴포넌트 제거
